Disable Docker updates rolling window on DEBUG level
