+++ sematic/plugins/building/BUILD
@@ -28,5 +28,6 @@ sematic_py_lib(
         "//sematic:container_images",
         "//sematic/plugins:abstract_builder",
         "//sematic/utils:env",
+        "//sematic/utils:spinner",
     ],
 )
